OLD/requirements.txt
pandas
transformers
torch
scikit-learn
matplotlib
plotly
dash
dash-bootstrap-components
networkx
pydantic
torch
sentence-transformers
transformers
faiss-cpu